backend/requirements.txt
alembic==1.9.4
asgiref==3.6.0
Authlib==1.1.0
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.9.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==21.0.0
azure-mgmt-subscription==3.0.0
Babel==2.11.0
bcrypt==3.2.2
bidict==0.22.1
blinker==1.5
boto3==1.26.76
botocore==1.29.76
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
click==8.1.3
cryptography==3.4.8
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.3.1
eventlet==0.33.0
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
httpagentparser==1.9.5
idna==2.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
ldap3==2.9.1
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==0.3.1
msrest==0.7.1
mysqlclient==2.1.1
oauthlib==3.2.2
paramiko==3.0.0
passlib==1.7.4
Pillow==9.4.0
portalocker==2.7.0
psutil==5.9.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyotp==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
qrcode==7.4.2
requests==2.25.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
simplejson==3.18.3
six==1.16.0
speaklater3==1.4
SQLAlchemy==1.4.46
sqlparse==0.4.3
sshtunnel==0.4.0
typing_extensions==4.5.0
ua-parser==0.16.1
urllib3==1.26.14
user-agents==2.2.0
Werkzeug==2.1.2
WTForms==3.0.1
